finished up the persistance of the cities
